Entities in progress

+++ src/main/java/ru/deloom/jpaservice/repositories/KeywordsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ru.deloom.jpaservice.entities.Keywords;
 
-public interface KeywordsRepository extends JpaRepository<Keywords, Integer>{
+public interface KeywordsRepository extends JpaRepository<Keywords, Long>{
 	List<Keywords> findByName(String name);
 }
